CAP-6 Reformatted Screen and Setup Navigation for future
