+++ frontend/30DaysChallenge/App.tsx
@@ -1,157 +1,15 @@
-import {
-  StyleSheet,
-  View,
-  Text,
-  TextInput,
-  SafeAreaView,
-  TouchableWithoutFeedback,
-  Keyboard,
-  TouchableOpacity,
-} from "react-native";
-import {
-  useFonts,
-  Inter_900Black,
-  Inter_800ExtraBold,
-  Inter_400Regular,
-} from "@expo-google-fonts/inter";
-import React, { useEffect, useCallback, ReactNode, useState } from "react";
-import * as SplashScreen from "expo-splash-screen";
-// import { Button } from "react-native-paper";
+import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import LandingPage from "./screens/LandingPage";
 
-SplashScreen.preventAutoHideAsync();
-
-interface Props {
-  children?: ReactNode;
-}
-const DismissKeyboard = ({ children }: Props) => (
-  <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
-    {children}
-  </TouchableWithoutFeedback>
-);
 export default function App() {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  let [fontsLoaded, error] = useFonts({
-    Inter_900Black,
-    Inter_800ExtraBold,
-    Inter_400Regular,
-  });
-
-  useEffect(() => {
-    async function prepare() {
-      await SplashScreen.preventAutoHideAsync();
-    }
-    prepare();
-  }, []);
-
-  const onLayoutRootView = useCallback(async () => {
-    if (fontsLoaded) {
-      await SplashScreen.hideAsync();
-    }
-  }, [fontsLoaded]);
-
-  if (!fontsLoaded) {
-    return null;
-  }
-
-  const handleLogin = () => {
-    // #TODO
-    // This is where we will use the backend to check the login values to see if they are valid
-
-    console.log("username: " + username +" password: "+ password, " this is the login information");
-  };
+  const Stack = createNativeStackNavigator();
   return (
-    <DismissKeyboard>
-      <SafeAreaView style={styles.container} onLayout={onLayoutRootView}>
-        <Text style={styles.text}>30 Days Challenge</Text>
-        <View style={styles.inputContainer}>
-          <TextInput
-            style={styles.input}
-            placeholder="Username"
-            value={username}
-            onChangeText={(val) => setUsername(val)}
-          />
-          <TextInput
-            style={styles.input}
-            placeholder="Password"
-            secureTextEntry={true}
-            value={password}
-            onChangeText={(val) => setPassword(val)}
-          />
-        </View>
-        <TouchableOpacity style={styles.button} onPress={handleLogin}>
-          <Text style={styles.buttonText}>Login</Text>
-        </TouchableOpacity>
-        <TouchableOpacity style={styles.register}>
-          <Text style={styles.registerText}>
-            New to 30 Days Challenge? Click here to register
-          </Text>
-        </TouchableOpacity>
-      </SafeAreaView>
-    </DismissKeyboard>
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="LandingPage">
+        <Stack.Screen name="LandingPage" component={LandingPage} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: "#123166",
-    width: "100%",
-    height: "100%",
-    display: "flex",
-  },
-  text: {
-    fontFamily: "Inter_800ExtraBold",
-    color: "white",
-    fontSize: 70,
-    textAlign: "center",
-    alignItems: "center",
-    width: "100%",
-    height: 300,
-    position: "absolute",
-    lineHeight: 105,
-    top: 95,
-  },
-  inputContainer: {},
-  input: {
-    height: 44,
-    width: 327,
-    fontFamily: "Inter_400Regular",
-    margin: 12,
-    borderWidth: 1,
-    padding: 10,
-    top: 439,
-    alignSelf: "center",
-    backgroundColor: "#FFFFFF",
-    borderRadius: 4,
-  },
-  button: {
-    backgroundColor: "#F45D9A",
-    width: 327,
-    height: 44,
-    top: 449,
-    alignSelf: "center",
-    borderRadius: 4,
-    justifyContent: "center",
-  },
-  buttonText: {
-    color: "#FFFFFF",
-    fontFamily: "Inter_400Regular",
-    alignSelf: "center",
-    fontSize: 17,
-  },
-  register: {
-    width: 327,
-    height: 48,
-    top: 459,
-    justifyContent: "center",
-    textAlign: "center",
-  },
-  registerText: {
-    color: "#FFFFFF",
-    fontFamily: "Inter_400Regular",
-    alignSelf: "center",
-    fontSize: 17,
-    textAlign: "center",
-    left: 35,
-  },
-});
