+++ frontend/30DaysChallenge/package.json
@@ -11,6 +11,9 @@
   "dependencies": {
     "@expo-google-fonts/inter": "^0.2.2",
     "@expo/webpack-config": "^0.17.2",
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/native": "^6.1.2",
+    "@react-navigation/native-stack": "^6.9.8",
     "expo": "~47.0.8",
     "expo-app-loading": "~2.1.1",
     "expo-font": "~11.0.1",
@@ -19,7 +22,11 @@
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
+    "react-native-gesture-handler": "~2.8.0",
     "react-native-paper": "5.0.0-rc.10",
+    "react-native-reanimated": "~2.12.0",
+    "react-native-safe-area-context": "4.4.1",
+    "react-native-screens": "~3.18.0",
     "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.18.9"
   },
